Add skeleton of code search guide
This is to see what the structure/layout of the doc looks like in the learn site. It copies over some but not all of the content from the WIP Google Doc at https://docs.google.com/document/d/1fw0ee634RwLGJgB4jHi049k5K2VL-q47JgOcZEc01IU/edit. That Google Doc will contain the latest content; this is just for layout previewing.

See https://github.com/sourcegraph/learn/issues/113 for a discussion of whether this is the right place for the code search guide to live.