Adding issue templates
We added a "Submit an idea" column on the [Handbook Marketing page](https://about.sourcegraph.com/handbook/marketing), and I am updating this so we can get meaningful ideas. 

While I'm here, I loaded in the Bug report and Feature request templates. I am open to feedback for each of these, and we should continue to iterate in any case