Light edits to the new sentence
From

```
The above command will recursively add all the modified files in all subdirectories in the current repository to the staging area, regardless of where we are running the command from. We can run this from any subdirectory of the repository.
```

to 

```
The above command will recursively add all the repository's modified files in all subdirectories to the staging area, regardless of where we are running the command. We can run this from any location within the repository.
```